+++ insar_explorer.py
@@ -239,6 +239,7 @@ class InsarExplorer:
         if click_handler is not None:
             click_handler.clearReferenceFeatureHighlight()
         if controller is not None:
+            controller.clearTimeSeriesClipboard()
             controller.removePolygonDrawingTool(reference=True)
 
         # print "** UNLOAD InsarExplorer"