+++ src/ui/time_series/action_icons.py
@@ -1,0 +1,5 @@
+"""Stable registered icons for time-series context-menu categories."""
+
+STYLE_ACTION_ICON = ":/icons/icons/plot_settings.svg"
+FIT_ACTION_ICON = ":/icons/icons/fit_curve.svg"
+REPLICA_ACTION_ICON = ":/icons/icons/replica.svg"